# ResourceServer/src/main/resources/static/swagger.yaml
openapi: "3.0.3"
info:
  title: "Twister Resource Server API"
  version: "0.0.1"
components:
  securitySchemes:
    oAuthSample:
      type: oauth2
      description: API использует [OAuth2 authorizationCode](https://auth0.com/docs/get-started/authentication-and-authorization-flow/authorization-code-flow).
      flows:
        authorizationCode:
          tokenUrl: https://twister.fvds.ru:8443/realms/twister/protocol/openid-connect/token
          authorizationUrl: https://twister.fvds.ru:8443/realms/twister/protocol/openid-connect/auth
          scopes: { }
  #            write: записать значение
  #            read: прочитать значение
  schemas:
    MoneyTypePartialResponseDto:
      type: "object"
      required:
        - name
        - path_to_icon
      properties:
        name:
          description: "Имя типа по которому его можно однозначно идентифицировать"
          type: "string"
          maxLength: 64
          minLength: 1
          example: "name"
        path_to_icon:
          description: "Путь до иконки"
          maxLength: 64
          minLength: 1
          type: "string"
          example: "icon"
...
    ItemMoneyPartialResponseDto:
       required:
          - cost
          - name
          - type
          - visible_name
       type: object
       properties:
         name:
           description: "Имя предмета по которому его можно однозначно идентифицировать"
           type: "string"
           maxLength: 64
           example: "name"
           minLength: 1
         visible_name:
           description: "Имя видимое пользователям"
           type: "string"
           example: "name"
           maxLength: 64
           minLength: 1
         type:
           description: "Тип предмета"
           type: "string"
           enum:
            - "MONEY"
         cost:
           $ref: "#/components/schemas/MoneyPartialResponseDto"
    ItemTrashPartialResponseDto:
        required:
          - name
          - type
          - visible_name
        type: object
        properties:
          name:
            description: "Имя предмета по которому его можно однозначно идентифицировать"
            type: "string"
            example: "name"
            maxLength: 64
            minLength: 1
          visible_name:
            description: "Имя видимое пользователям"
            type: "string"
            example: "name"
            maxLength: 64
            minLength: 1
          type:
            description: "Тип предмета"
            type: "string"
            enum:
              - "TRASH"
    MoneyPartialResponseDto:
      type: "object"
      required:
        - value
        - type
      properties:
        type:
          $ref: "#/components/schemas/MoneyTypePartialResponseDto"
        value:
          description: "Количество денег"
          minimum: 0
          example: 0
          type: number
...
    MoneyCreateRequestDto:
      type: "object"
      required:
        - type
        - value
      properties:
        type:
          $ref: "#/components/schemas/MoneyTypeNameRequestDto"
        value:
          description: "Количество денег"
          minimum: 0
          example: 0
          type: number
    MoneyTypeNameRequestDto:
      type: "object"
      required:
        - name
      properties:
        name:
          description: "Имя типа по которому его можно однозначно идентифицировать"
          type: "string"
          maxLength: 64
          minLength: 1
          example: "name"
    ItemPartialResponseDto:
      oneOf:
        - $ref: "#/components/schemas/ItemMoneyPartialResponseDto"
        - $ref: "#/components/schemas/ItemTrashPartialResponseDto"
      discriminator:
        propertyName: type
        mapping:
          MONEY: "#/components/schemas/ItemMoneyPartialResponseDto"
          TRASH: "#/components/schemas/ItemTrashPartialResponseDto"
...
    ItemNameRequestDto:
      type: "object"
      required:
        - name
      properties:
        name:
          description: "Имя предмета по которому его можно однозначно идентифицировать"
          type: "string"
          example: "name"
          maxLength: 64
          minLength: 1
    CaseSlotCreateRequestDto:
      type: "object"
      required:
        - win_rate
        - item
        - quantity
      properties:
        win_rate:
          description: "Шанс выпадения предмета"
          minimum: 0
          maximum: 100
          example: 100
          type: number
        item:
          $ref: "#/components/schemas/ItemNameRequestDto"
        quantity:
          description: "Количество предметов в слоте"
          minimum: 0
          example: 1
          type: "integer"
          format: "int32"
    CasePartialResponseDto:
      type: "object"
      required:
        - price
        - cooldown
        - name
        - visible_name
        - description
      properties:
        price:
          $ref: "#/components/schemas/MoneyPartialResponseDto"
        cooldown:
          type: string
          format: iso8601
          description: "Кулдаун с последней прокрутки пользователя"
          example: P1S
        name:
          description: "Имя кейса по которому его можно однозначно идентифицировать"
          type: "string"
          maxLength: 64
          minLength: 1
          example: "name"
        visible_name:
          description: "Имя видимое пользователю"
          type: "string"
          maxLength: 64
          minLength: 1
          example: "name"
        description:
          description: "Описание кейса"
          type: "string"
          maxLength: 1000
          minLength: 0
          example: ""
    CaseCreateRequestDto:
      type: "object"
      required:
        - price
        - cooldown
        - visible_name
        - description
      properties:
        items:
          description: "**В сумме `win_rate` должен быть `100`**"
          type: "array"
          items:
            $ref: "#/components/schemas/CaseSlotCreateRequestDto"
        price:
          $ref: "#/components/schemas/MoneyCreateRequestDto"
        cooldown:
          type: string
          format: iso8601
          description: "Кулдаун с последней прокрутки пользователя"
          example: P1S
        visible_name:
          description: "Имя видимое пользователю"
          type: "string"
          maxLength: 64
          minLength: 1
          example: "name"
        description:
          description: "Описание кейса"
          type: "string"
          maxLength: 1000
          minLength: 0
          example: ""
    CaseSlotPartialResponseDto:
      type: "object"
      required:
        - win_rate
        - item
        - quantity
        - name
      properties:
        win_rate:
          description: "Шанс выпадения предмета"
          minimum: 0
          maximum: 100
          type: number
        item:
          $ref: "#/components/schemas/ItemPartialResponseDto"
        quantity:
          description: "Количество предметов в слоте"
          minimum: 0
          type: "integer"
          format: "int32"
        name:
          description: "Имя слота по которому его можно однозначно идентифицировать"
          type: "string"
          maxLength: 64
          minLength: 1
    CaseItemsPartialResponseDto:
      type: "object"
      required:
        - price
        - cooldown
        - visible_name
        - description
        - name
        - items
      properties:
        items:
          description: "**В сумме `win_rate` должен быть `100`**"
          type: "array"
          items:
            $ref: "#/components/schemas/CaseSlotPartialResponseDto"
        price:
          $ref: "#/components/schemas/MoneyPartialResponseDto"
        cooldown:
          type: string
          format: iso8601
          description: "Кулдаун с последней прокрутки пользователя"
          example: P1S
        name:
          description: "Имя кейса по которому его можно однозначно идентифицировать"
          type: "string"
          maxLength: 64
          minLength: 1
        visible_name:
          description: "Имя видимое пользователю"
          type: "string"
          maxLength: 64
          minLength: 1
        description:
          description: "Описание кейса"
          type: "string"
          maxLength: 1000
          minLength: 1
    CaseItemsWithoutCooldownPartialResponseDto:
      type: "object"
      required:
        - price
        - visible_name
        - description
        - name
        - items
      properties:
        items:
          description: "**В сумме `win_rate` должен быть `100`**"
          type: "array"
          items:
            $ref: "#/components/schemas/CaseSlotPartialResponseDto"
        price:
          $ref: "#/components/schemas/MoneyPartialResponseDto"
        name:
          description: "Имя кейса по которому его можно однозначно идентифицировать"
          type: "string"
          maxLength: 64
          minLength: 1
        visible_name:
          description: "Имя видимое пользователю"
          type: "string"
          maxLength: 64
          minLength: 1
        description:
          description: "Описание кейса"
          type: "string"
          maxLength: 1000
          minLength: 1
    CaseLightWithoutCooldownPartialResponseDto:
      type: "object"
      required:
        - price
        - visible_name
        - description
        - name
      properties:
        price:
          $ref: "#/components/schemas/MoneyPartialResponseDto"
        name:
          description: "Имя кейса по которому его можно однозначно идентифицировать"
          type: "string"
          maxLength: 64
          minLength: 1
        visible_name:
          description: "Имя видимое пользователю"
          type: "string"
          maxLength: 64
          minLength: 1
        description:
          description: "Описание кейса"
          type: "string"
          maxLength: 1000
          minLength: 1
    CaseWithoutCooldownPagePartialResponseDto:
      type: "object"
      required:
        - values
        - total_pages
        - page
      properties:
        values:
          type: "array"
          items:
            $ref: "#/components/schemas/CaseLightWithoutCooldownPartialResponseDto"
        total_pages:
          description: "Полное количество страниц кейсов"
          type: "integer"
          format: "int32"
          minimum: 0
          example: 1
        page:
          description: "Номер страницы из которой взята информация. Индексация начиная с нуля"
          type: "integer"
          format: "int32"
          minimum: 0
          example: 0
    CaseLightPartialResponseDto:
      type: "object"
      required:
        - price
        - cooldown
        - name
        - visible_name
        - description
      properties:
        price:
          $ref: "#/components/schemas/MoneyPartialResponseDto"
        cooldown:
          type: string
          format: iso8601
          description: "Кулдаун с последней прокрутки пользователя"
          example: P0S
        name:
          description: "Имя кейса по которому его можно однозначно идентифицировать"
          type: "string"
          maxLength: 64
          minLength: 1
          example: "name"
        visible_name:
          description: "Имя видимое пользователю"
          type: "string"
          maxLength: 64
          minLength: 1
          example: "name"
        description:
          description: "Описание кейса"
          type: "string"
          maxLength: 1000
          minLength: 0
          example: ""
    CasePagePartialResponseDto:
      type: "object"
      required:
        - values
        - total_pages
        - page
      properties:
        values:
          type: "array"
          items:
            $ref: "#/components/schemas/CaseLightPartialResponseDto"
        total_pages:
          description: "Полное количество страниц кейсов"
          type: "integer"
          format: "int32"
          minimum: 0
          example: 1
        page:
          description: "Номер страницы из которой взята информация. Индексация начиная с нуля"
          type: "integer"
          format: "int32"
          minimum: 0
          example: 0
...
    NotFoundException:
      type: "object"
      required:
        - message
      properties:
        message:
          description: "Не найден"
          type: "string"
          example: "Not found."
...
security:
  - oAuthSample:
      - write
      - read
paths:
...
  /api/case/{name}:
    put:
      tags:
        - case
      security:
        - oAuthSample:
            - read
            - write
      summary: "PUT api/case/{name}"
      description: "Поместить новый или существующий кейс.<br>На данный момент невозможно изменение слотов.<br>**Это API доступно только для роли `ADMIN`**"
      operationId: "putCase"
      parameters:
        - name: "name"
          in: "path"
          required: true
          description: "Имя кейса по которому его можно однозначно идентифицировать"
          schema:
            type: "string"
            maxLength: 64
            minLength: 1
        - name: "content-type"
          in: "header"
          required: false
          schema:
            type: "string"
            enum:
              - "application/json"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseCreateRequestDto"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/CasePartialResponseDto"
    get:
      tags:
        - case
      security:
        - oAuthSample:
            - read
      summary: "GET api/case/{name}"
      description: "Получить полную информацию по конкретному кейсу<br>**Это API доступно только для роли `ADMIN`**"
      operationId: "info"
      parameters:
        - name: "name"
          in: "path"
          required: true
          schema:
            description: "Имя кейса по которому его можно однозначно идентифицировать"
            type: "string"
            maxLength: 64
            minLength: 1
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/CaseItemsPartialResponseDto"
        "404":
          description: "Тип не найден"
          content:
            'application/json':
              schema:
                oneOf:
                  - $ref: "#/components/schemas/NotFoundException"
  /api/public/case:
    get:
      description: "Получить страницу кейсов"
      security: []
      tags:
        - public case
      summary: "GET api/public/case"
      operationId: "listCase"
      parameters:
        - name: "page"
          in: "query"
          required: false
          description: "номер страницы"
          schema:
            type: "integer"
            format: "int32"
            minimum: 0
            example: 0
            default: 0
        - name: "size"
          in: "query"
          required: false
          description: "То сколько будет кейсов в каждой странице"
          schema:
            type: "integer"
            format: "int32"
            minimum: 1
            maximum: 50
            example: 5
            default: 5
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/CaseWithoutCooldownPagePartialResponseDto"
  /api/public/case/{name}:
    get:
      description: "Получить информацию об конкретном кейсе"
      security: []
      tags:
        - public case
      summary: "GET api/public/case/{name}"
      operationId: "info"
      parameters:
        - name: "name"
          in: "path"
          required: true
          description: "Имя кейса по которому его можно однозначно идентифицировать"
          schema:
            type: "string"
            maxLength: 64
            minLength: 1
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/CaseItemsWithoutCooldownPartialResponseDto"
        "404":
          description: "Кейс не найден"
          content:
            'application/json':
              schema:
                oneOf:
                  - $ref: "#/components/schemas/NotFoundException"
  /api/general/{name}/actual_case:
    get:
      description: "Получить страницу кейсов с информацией о том сколько осталось времени, чтобы покрутить еще раз. Если прокрутки никогда не было, то `0`."
      tags:
        - actual case
      security:
        - oAuthSample:
            - read
      summary: "GET api/general/{name}/actual_case"
      operationId: "listCase"
      parameters:
        - name: "name"
          in: "path"
          required: true
          description: "Имя аккаунта по которому его можно однозначно идентифицировать"
          schema:
            type: "string"
            maxLength: 64
            minLength: 1
        - name: "page"
          in: "query"
          required: false
          description: "номер страницы"
          schema:
            type: "integer"
            format: "int32"
            minimum: 0
            example: 0
            default: 0
        - name: "size"
          in: "query"
          required: false
          description: "То сколько будет кейсов в каждой странице"
          schema:
            type: "integer"
            format: "int32"
            minimum: 1
            maximum: 50
            example: 5
            default: 5
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/CasePagePartialResponseDto"
        "404":
          description: "Аккаунт не найден"
          content:
            'application/json':
              schema:
                oneOf:
                  - $ref: "#/components/schemas/NotFoundException"
...